fixed overwrite system
The default is True again and this will cause files to be overwritten, but folders will not be overwritten. To overwrite a folder it can be done manually as generally people shouldn't want to do this.
Also corrected some issues with the merging of scan.tsv files and the ordering of the columns (bids-validator was giving me an error about no filenames column since it was being put second, so now they are explicitly put in the correct order)